# docs/swagger/dcc-server-openapi.yaml
openapi: 3.0.3
info:
  title: DCC Server API
  version: 1.0.0
  description: |
    Full API surface for DCC Express server, including `/api/*` application routes
    and OpenAI-compatible facade routes under `/v1/*`.
servers:
  - url: /
    description: Current host (recommended)
  - url: http://localhost:3000
    description: Local default server
...
paths:
...
  /api/settings:
    get:
      tags: [Settings]
      summary: Get app settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  maxRecommendedDefinitions: { type: integer, minimum: 3, maximum: 8 }
                  geminiApiKey: { type: string }
                  geminiClient: { type: string }
                  geminiModel: { type: string }
                  geminiConnectorId: { type: string }
                  geminiConnectorBaseUrl: { type: string }
                  geminiConnectorApiKey: { type: string }
                  geminiConnectorModel: { type: string }
                  geminiConnectorMode: { type: string, enum: [regular, raw] }
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Settings]
      summary: Update app settings
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                repoUrl: { type: string }
                repoPath: { type: string }
                maxRecommendedDefinitions: { type: integer }
                geminiClient: { type: string }
                geminiApiKey: { type: string }
                geminiModel: { type: string }
                geminiConnectorId: { type: string }
                geminiConnectorBaseUrl: { type: string }
                geminiConnectorApiKey: { type: string }
                geminiConnectorModel: { type: string }
                geminiConnectorMode: { type: string, enum: [regular, raw] }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "500":
          $ref: "#/components/responses/ServerError"
  /api/asset-repos:
    get:
      tags: [Asset Repos]
      summary: List asset repositories
      responses:
        "200":
          description: Repositories
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AssetRepo" }
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Asset Repos]
      summary: Create asset repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetRepoCreate"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetRepo"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/asset-repos/{id}:
    put:
      tags: [Asset Repos]
      summary: Update asset repository
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssetRepoUpdate"
      responses:
        "200":
          description: Updated repo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssetRepo"
        "404":
          $ref: "#/components/responses/NotFound"
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
    delete:
      tags: [Asset Repos]
      summary: Delete asset repository
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /api/onboarding-status:
    get:
      tags: [Onboarding]
      summary: Get onboarding seen status
      responses:
        "200":
          description: Onboarding status
          content:
            application/json:
              schema:
                type: object
                properties:
                  seen: { type: boolean }
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Onboarding]
      summary: Set onboarding seen status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seen: { type: boolean }
      responses:
        "200":
          description: Updated status
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  seen: { type: boolean }
        "500":
          $ref: "#/components/responses/ServerError"
  /api/current-dev-project:
    get:
      tags: [Settings]
      summary: Get current dev project
      responses:
        "200":
          description: Current dev project path
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string }
        "500":
          $ref: "#/components/responses/ServerError"
    post:
      tags: [Settings]
      summary: Set current dev project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path: { type: string }
              required: [path]
      responses:
        "200":
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  path: { type: string }
        "500":
          $ref: "#/components/responses/ServerError"
  /api/current-dev-project/context-window:
    get:
      tags: [Settings]
      summary: Resolve context window token size for the current or specified project
      parameters:
        - name: path
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Context window result
          content:
            application/json:
              schema:
                type: object
                properties:
                  path: { type: string }
                  contextWindowTokens: { type: integer, nullable: true }
        "500":
          $ref: "#/components/responses/ServerError"
  /api/settings/export:
    post:
      tags: [Settings]
      summary: Export settings and related records to a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath: { type: string }
              required: [filePath]
      responses:
        "200":
          description: Export completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  filePath: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/settings/import:
    post:
      tags: [Settings]
      summary: Import settings and related records from a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath: { type: string }
              required: [filePath]
      responses:
        "200":
          description: Import completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  filePath: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/database/backup:
    post:
      tags: [Settings]
      summary: Backup SQLite database to a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath: { type: string }
              required: [filePath]
      responses:
        "200":
          description: Backup completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  filePath: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /api/database/restore:
    post:
      tags: [Settings]
      summary: Restore SQLite database from a file
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filePath: { type: string }
              required: [filePath]
      responses:
        "200":
          description: Restore completed
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  filePath: { type: string }
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
...
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: integer }
  schemas:
    ErrorEnvelope:
      type: object
      properties:
        error: { type: string }
...
    AssetRepo:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        remoteUrl: { type: string }
        localPath: { type: string }
        enabled: { type: boolean }
        createdAt: { type: string, nullable: true }
        updatedAt: { type: string, nullable: true }
    AssetRepoCreate:
      type: object
      properties:
        name: { type: string }
        remoteUrl: { type: string }
        localPath: { type: string }
        enabled: { type: boolean }
      required: [name, remoteUrl, localPath]
    AssetRepoUpdate:
      type: object
      properties:
        name: { type: string }
        remoteUrl: { type: string }
        localPath: { type: string }
        enabled: { type: boolean }
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"